server/requirements.txt
pandas
scikit-learn
pymongo
python-dotenv
fastapi
uvicorn
passlib
bcrypt
python-jose
python-multipart
json
requests